Added initial version demo.
The demo folder with old LightStone demo (v0.2) added.  This demo is
meant only to show that the engine works and is not a good
representation of what can be done.  Much more could be accomplished
using this engine. Game runs slow when full screen for some reason. I
have not found a decent fix for the framerate drop outside of using
Adobe Air or porting to another language.  However, it runs very smooth
at its normal web-friendly resolution.