Demo working update.
The demo supporting files were not added in previous sync, and the demo
was non-functional.  This update adds the required files and makes sure
the demo is the working Air compiled version rather than the
non-functional Flash SWF version.  Run the exe file named LightStone.exe
in the LightStone folder to see what I got done before opening the
project up to everyone in the form of an open source RPG framework.
Level data used in Ogmo.